move asm related helper functions from lib.rs to utils.rs
